src/services/telegram_service.py
"""
Telegram Bot Integration Service (Placeholder).

This service provides integration with Telegram for receiving documents
and triggering the document processing pipeline.

NOTE: This is a placeholder implementation. Full implementation requires:
1. Telegram Bot Token from @BotFather
2. Webhook setup or polling configuration
3. Production deployment with HTTPS for webhooks
"""

import os
from typing import Optional, Dict, Any, List, Callable, Awaitable
from enum import Enum
from pydantic import BaseModel, Field
from datetime import datetime

# Telegram imports will be used when fully implemented
try:
    from telegram import Update, Bot, Document as TelegramDocument
    from telegram.ext import Application, CommandHandler, MessageHandler, filters
    TELEGRAM_AVAILABLE = True
except ImportError:
    TELEGRAM_AVAILABLE = False


class TelegramMessageType(str, Enum):
    """Types of Telegram messages."""
    TEXT = "text"
    DOCUMENT = "document"
    PHOTO = "photo"
    COMMAND = "command"


class IncomingDocument(BaseModel):
    """Represents a document received from Telegram."""
    file_id: str = Field(description="Telegram file ID")
    file_name: str = Field(description="Original filename")
    file_size: int = Field(description="File size in bytes")
    mime_type: Optional[str] = Field(default=None, description="MIME type if detected")
    sender_id: int = Field(description="Telegram user ID of sender")
    sender_username: Optional[str] = Field(default=None, description="Telegram username")
    chat_id: int = Field(description="Chat ID where document was sent")
    received_at: datetime = Field(default_factory=datetime.utcnow)
    caption: Optional[str] = Field(default=None, description="Document caption if provided")


class TelegramBotConfig(BaseModel):
    """Configuration for Telegram bot."""
    bot_token: str = Field(description="Telegram Bot API token")
    webhook_url: Optional[str] = Field(default=None, description="Webhook URL for receiving updates")
    allowed_users: List[int] = Field(default_factory=list, description="List of allowed user IDs (empty = all)")
    allowed_chats: List[int] = Field(default_factory=list, description="List of allowed chat IDs (empty = all)")
    max_file_size_mb: int = Field(default=20, description="Maximum file size in MB")


class TelegramServiceStatus(BaseModel):
    """Status of the Telegram service."""
    enabled: bool
    connected: bool
    bot_username: Optional[str] = None
    webhook_configured: bool = False
    last_update_id: Optional[int] = None
    message: str = ""


class TelegramService:
    """
    Placeholder Telegram Bot Service.
    
    This service will handle:
    1. Receiving documents from Telegram
    2. Downloading files from Telegram servers
    3. Triggering document processing pipeline
    4. Sending status updates back to users
    """
    
    def __init__(self):
        self.bot_token = os.getenv("TELEGRAM_BOT_TOKEN")
        self.webhook_url = os.getenv("TELEGRAM_WEBHOOK_URL")
        self.enabled = bool(self.bot_token) and TELEGRAM_AVAILABLE
        self.bot: Optional[Any] = None
        self.application: Optional[Any] = None
        self._document_handler: Optional[Callable[[IncomingDocument, bytes], Awaitable[None]]] = None
        
        # Placeholder for allowed users/chats (configure via environment or database)
        self.allowed_users: List[int] = []
        self.allowed_chats: List[int] = []
    
    def get_status(self) -> TelegramServiceStatus:
        """Get current status of Telegram service."""
        return TelegramServiceStatus(
            enabled=self.enabled,
            connected=self.bot is not None,
            bot_username=None,  # Would be populated when connected
            webhook_configured=bool(self.webhook_url),
            message="Telegram integration is a placeholder - configure TELEGRAM_BOT_TOKEN to enable"
        )
    
    def set_document_handler(
        self, 
        handler: Callable[[IncomingDocument, bytes], Awaitable[None]]
    ):
        """
        Set the handler function for incoming documents.
        
        The handler will be called with:
        - IncomingDocument: Metadata about the received document
        - bytes: The actual file content
        """
        self._document_handler = handler
    
    async def initialize(self) -> bool:
        """
        Initialize the Telegram bot.
        
        Returns:
            True if initialization was successful
        """
        if not self.enabled:
            print("[Telegram] Service disabled - TELEGRAM_BOT_TOKEN not configured")
            return False
        
        try:
            # Placeholder: Would initialize the bot here
            # self.application = Application.builder().token(self.bot_token).build()
            # self.bot = self.application.bot
            # 
            # # Add handlers
            # self.application.add_handler(CommandHandler("start", self._handle_start))
            # self.application.add_handler(CommandHandler("help", self._handle_help))
            # self.application.add_handler(CommandHandler("status", self._handle_status))
            # self.application.add_handler(
            #     MessageHandler(filters.Document.ALL, self._handle_document)
            # )
            
            print("[Telegram] Bot initialized (placeholder)")
            return True
            
        except Exception as e:
            print(f"[Telegram] Failed to initialize: {e}")
            return False
    
    async def start_polling(self):
        """Start polling for updates (for development/testing)."""
        if not self.enabled:
            print("[Telegram] Cannot start polling - service disabled")
            return
        
        # Placeholder: Would start polling here
        # await self.application.run_polling()
        print("[Telegram] Polling started (placeholder)")
    
    async def setup_webhook(self, webhook_url: str) -> bool:
        """
        Set up webhook for receiving updates.
        
        Args:
            webhook_url: HTTPS URL to receive updates
        
        Returns:
            True if webhook was set up successfully
        """
        if not self.enabled:
            return False
        
        # Placeholder: Would set up webhook here
        # await self.bot.set_webhook(url=webhook_url)
        print(f"[Telegram] Webhook configured: {webhook_url} (placeholder)")
        return True
    
    async def process_webhook_update(self, update_data: Dict[str, Any]) -> bool:
        """
        Process an update received via webhook.
        
        Args:
            update_data: Raw update data from Telegram
        
        Returns:
            True if update was processed successfully
        """
        if not self.enabled:
            return False
        
        # Placeholder: Would process update here
        # update = Update.de_json(update_data, self.bot)
        # await self.application.process_update(update)
        print(f"[Telegram] Processing webhook update (placeholder)")
        return True
    
    async def _handle_start(self, update: Any, context: Any):
        """Handle /start command."""
        # Placeholder implementation
        pass
    
    async def _handle_help(self, update: Any, context: Any):
        """Handle /help command."""
        help_text = """
🤖 *IDP Bot - Intelligent Document Processing*

I can process documents and route them to the appropriate departments.

*Commands:*
/start - Start the bot
/help - Show this help message
/status - Check processing status

*Supported Documents:*
📄 PDF, Word, Excel, PowerPoint
📐 CAD files (DWG, DXF, IFC)
📊 Spreadsheets (XLSX, CSV)

Just send me a document and I'll process it!
        """
        # await update.message.reply_text(help_text, parse_mode='Markdown')
        pass
    
    async def _handle_status(self, update: Any, context: Any):
        """Handle /status command."""
        # Placeholder: Would check processing status here
        pass
    
    async def _handle_document(self, update: Any, context: Any):
        """Handle incoming documents."""
        if not self._document_handler:
            print("[Telegram] No document handler configured")
            return
        
        # Placeholder: Would download and process document here
        # document = update.message.document
        # file = await context.bot.get_file(document.file_id)
        # file_content = await file.download_as_bytearray()
        # 
        # incoming_doc = IncomingDocument(
        #     file_id=document.file_id,
        #     file_name=document.file_name,
        #     file_size=document.file_size,
        #     mime_type=document.mime_type,
        #     sender_id=update.message.from_user.id,
        #     sender_username=update.message.from_user.username,
        #     chat_id=update.message.chat_id,
        #     caption=update.message.caption
        # )
        # 
        # await self._document_handler(incoming_doc, bytes(file_content))
        pass
    
    async def send_message(self, chat_id: int, text: str, parse_mode: str = "Markdown") -> bool:
        """
        Send a message to a chat.
        
        Args:
            chat_id: Target chat ID
            text: Message text
            parse_mode: Telegram parse mode (Markdown, HTML)
        
        Returns:
            True if message was sent successfully
        """
        if not self.enabled:
            print(f"[Telegram] Would send to {chat_id}: {text}")
            return True
        
        # Placeholder: Would send message here
        # await self.bot.send_message(chat_id=chat_id, text=text, parse_mode=parse_mode)
        return True
    
    async def send_document_processed_notification(
        self,
        chat_id: int,
        document_name: str,
        departments: List[str],
        summary: str
    ) -> bool:
        """
        Send notification about processed document.
        
        Args:
            chat_id: Target chat ID
            document_name: Name of processed document
            departments: Departments document was routed to
            summary: Brief summary of document
        
        Returns:
            True if notification was sent
        """
        dept_list = ", ".join(departments)
        message = f"""
✅ *Document Processed Successfully*

📄 *Document:* {document_name}
🏢 *Routed to:* {dept_list}

📝 *Summary:*
{summary}
        """
        return await self.send_message(chat_id, message)


# Create singleton instance
telegram_service = TelegramService()


# API functions for webhook router
async def handle_telegram_webhook(update_data: Dict[str, Any]) -> bool:
    """
    Handle incoming Telegram webhook.
    
    This function should be called from the webhook router.
    """
    return await telegram_service.process_webhook_update(update_data)


def get_telegram_status() -> TelegramServiceStatus:
    """Get current Telegram service status."""
    return telegram_service.get_status()


def configure_document_handler(
    handler: Callable[[IncomingDocument, bytes], Awaitable[None]]
):
    """Configure the handler for incoming documents."""
    telegram_service.set_document_handler(handler)
