pyproject.toml
[project]
name = "fast-api-agent"
version = "0.1.0"
description = "Intelligent Document Processing for Infrastructure Operations"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiofiles>=25.1.0",
    "chromadb>=1.3.5",
    "docx2txt>=0.9",
    "fastapi[standard]>=0.121.1",
    "langchain-groq>=1.0.0",
    "langchain-mcp-adapters>=0.1.12",
    "langchain[google-genai]>=1.0.5",
    "langgraph>=0.4.8",
    "langgraph-supervisor>=0.0.10",
    "llama-index>=0.14.9",
    "llama-index-embeddings-huggingface>=0.6.1",
    "llama-index-readers-file>=0.5.5",
    "llama-index-vector-stores-chroma>=0.5.3",
    "openpyxl>=3.1.5",
    "pydantic-settings>=2.12.0",
    "pypdf>=6.4.0",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.20",
    "python-pptx>=1.0.2",
    "python-telegram-bot>=21.0",
    "python-magic-bin>=0.4.14",
    "sendgrid>=6.11.0",
    "uvicorn>=0.38.0",
]
